refactor(storage): deepCreateFolders() -> fileHandlers.findOrCreateFolders()
